backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.25
requests==2.31.0
python-dotenv==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
redis==5.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-cov==4.1.0
python-jose==3.3.0
web3==6.11.4
